Separate export generation functions
